Add a newer verison of the password generation algorithm.
This one is more likely to get past password strength requirements,
and thus generates stronger passwords. The core of the algorithm is
still the same: concatenate the master password with a site name and
run a cryptographic hash on it. The differences are:

    1. Use SHA256 instead of MD5, to get longer outputs.
    2. Use Base64 to turn the hash into a string, instead of hex, to
       get a wider variety of characters.
    3. Append one known lower-case, upper-case, and number to the end
       of the password, derived from the hash, just to make sure.
    4. Append an exclamation point, to be excited, and also to pass
       the "symbol" requirement.

There is now a check box to determine what version to use; by default
we use the newer version. If you need the older version, check the
box. The program will remember what version that it last used for a
site, so that you don't have to keep checking the box for sites that
used the old algorithm.